+++ isitfit/cost/utilizationListener.py
@@ -10,12 +10,13 @@ from termcolor import colored
 
 class UtilizationListener:
 
-  def __init__(self):
+  def __init__(self, emailTo):
     # iterate over all ec2 instances
     self.sum_capacity = 0
     self.sum_used = 0
     self.df_all = []
     self.table = None # will contain the final table after calling `after_all`
+    self.emailTo = emailTo
 
 
   def per_ec2(self, ec2_obj, ec2_df, mm, ddg_df):
@@ -65,23 +66,49 @@ class UtilizationListener:
     dt_end   = mm.EndTime.strftime("%Y-%m-%d")
     
     self.table = [
-      ["Start date", "%s"%dt_start],
-      ["End date", "%s"%dt_end],
-      ["EC2 machines (total)", "%i"%n_ec2_total],
-      ["EC2 machines (analysed)", "%i"%n_ec2_analysed],
-      [colored("Billed cost", 'cyan'), colored("%0.0f $"%self.sum_capacity, 'cyan')],
-      [colored("Used cost", 'cyan'), colored("%0.0f $"%self.sum_used, 'cyan')],
-      [colored("CWAU (Used/Billed)", cwau_color), colored("%0.0f %%"%cwau_val, cwau_color)],
+            {'color': None,       'label': "Start date",              'value': "%s"%dt_start                },
+            {'color': None,       'label': "End date",                'value': "%s"%dt_end                  },
+            {'color': None,       'label': "EC2 machines (total)",    'value': "%i"%n_ec2_total             },
+            {'color': None,       'label': "EC2 machines (analysed)", 'value': "%i"%n_ec2_analysed          },
+            {'color': 'cyan',     'label': "Billed cost",             'value': "%0.0f $"%self.sum_capacity  },
+            {'color': 'cyan',     'label': "Used cost",               'value': "%0.0f $"%self.sum_used      },
+            {'color': cwau_color, 'label': "CWAU (Used/Billed)",      'value': "%0.0f %%"%cwau_val          },
     ]
 
 
   def display_all(self, *args, **kwargs):
+    def get_row(row):
+        def get_cell(i):
+          retc = row[i] if row['color'] is None else colored(row[i], row['color'])
+          return retc
+        
+        retr = [get_cell('label'), get_cell('value')]
+        return retr
+
+    dis_tab = [get_row(row) for row in self.table]
+
     # logger.info("Summary:")
     logger.info("Cost-Weighted Average Utilization (CWAU) of the AWS EC2 account:")
     logger.info("")
-    logger.info(tabulate(self.table, headers=['Field', 'Value']))
+    logger.info(tabulate(dis_tab, headers=['Field', 'Value']))
     logger.info("")
     logger.info("For reference:")
     logger.info(colored("* CWAU >= 70% is well optimized", 'green'))
     logger.info(colored("* CWAU <= 30% is underused", 'red'))
+
+  def share_email(self, *args, **kwargs):
+      # check if email requested
+      if self.emailTo is None:
+          return
+
+      if len(self.emailTo)==0:
+          return
+
+      from ..emailMan import EmailMan
+      em = EmailMan(
+        dataType='cost analyze',
+        dataVal={'table': self.table}
+      )
+      em.send(self.emailTo)
+